scraper/requirements.txt
# ============================================================
#  Job Scraper — Python Dependencies
#  Install with: pip install -r requirements.txt
#  Then run:     playwright install chromium
# ============================================================

# ── Web Scraping ─────────────────────────────────────────────
playwright==1.44.0
linkedin-api==2.3.0          # tomquirk/linkedin-api (unofficial)

# ── CV Parsing ───────────────────────────────────────────────
pdfplumber==0.11.0

# ── AI / Semantic Matching ───────────────────────────────────
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
# torch is a transitive dep of sentence-transformers;
# the CPU-only build is enough and avoids a huge download in GitHub Actions
torch==2.3.1+cpu ; sys_platform != "darwin"
torch==2.3.1     ; sys_platform == "darwin"

# ── Data Modelling ───────────────────────────────────────────
pydantic==2.7.4

# ── Database (Supabase) ──────────────────────────────────────
supabase==2.5.3

# ── Configuration ────────────────────────────────────────────
pyyaml==6.0.1

# ── Email Notifications ──────────────────────────────────────
resend==2.3.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1         # for local .env support outside GitHub Actions
httpx==0.27.0                # used internally by supabase client
